added simple counter to helloworld
